requirements.txt
langchain
langchain-community
langchain_experimental

google-generativeai
langchain-google-genai

streamlit 
streamlit_chat